requirements.txt
# Telegram Bot
python-telegram-bot==20.7
httpx==0.25.2

# HTTP客户端
aiohttp==3.9.1

# 数据处理
pandas==2.1.4
numpy==1.26.2

# 配置管理
PyYAML==6.0.1
pydantic==2.5.2
pydantic-settings==2.1.0

# 模板引擎
Jinja2==3.1.2

# 数据库
aiosqlite==0.19.0

# 日志
loguru==0.7.2

# 环境变量
python-dotenv==1.0.0

# 异步支持
asyncio==3.4.3

# Telegram Client API（用于转发服务，接收Bot消息）
pyrogram==2.0.106
tgcrypto==1.2.5